Still a work in progress.
A rewrite with a change in architecture. There's one common interface for collectors and analyzers and detectors which can be run in seprate threads. We now collect package information along with permissions, processor information, memory and network information. We also collect screen toggle events.
The analyzers then create threat models which the detectors use to figure out if a current activity is an actual threat or not. An alert can be issued for correct cases. The typical workflow is an analyzer runs and uses the information collected by the collectors and then a detector runs after it and uses the created model.

The current models don't seem to detect privacy-stealing malware in any useful way.